machine: Move file descriptor utility functions to filesystem
