+++ machine/include/filesystem.h
@@ -3,6 +3,7 @@
 
 #include "config.h"
 
+#include <stdbool.h>
 #include <stddef.h>
 
 typedef struct KFILE {
@@ -19,5 +20,14 @@ typedef struct FILEDESC {
 extern const KFILE files[];
 
 const KFILE* find_file(const char* filename);
+
+// File descriptor util functions
+bool fd_is_stdio(int fd);
+bool fd_is_valid(int fd, size_t num_fds);
+FILEDESC* get_fd_entry(int fd, FILEDESC* open_files, size_t num_fds);
+bool fd_entry_is_opened(FILEDESC* entry);
+bool fd_is_opened(int fd, FILEDESC* open_files, size_t num_fds);
+
+#define OPEN_FILE_FD_OFFSET 3
 
 #endif /* SBI_FILES_BASE */
